Move json_version into api2.shared
To break a cyclic import in api2.events.
