+++ securedrop/journalist_app/api2/shared.py
@@ -3,10 +3,14 @@ This module contains helper functions factored out of the v1 Journalist API
 (journalist_app.api) and shared with the v2 Journalist API (journalist_app.api2)
 """
 
+import hashlib
 import os
+from typing import Mapping
 from uuid import UUID
 
 from db import db
+from flask import json
+from journalist_app.api2.types import Version
 from journalist_app.sessions import session
 from models import (
     InvalidUUID,
@@ -16,6 +20,20 @@ from models import (
 )
 from sqlalchemy.exc import IntegrityError
 from store import Storage
+
+
+def json_version(d: Mapping) -> Version:
+    """
+    Calculate the version (BLAKE2s digest) of the normalized JSON representation
+    of the dictionary ``d``.
+
+    We use BLAKE2s here because SHA-256 is too slow (we don't care about
+    cryptographic security) and CRC-32 is too collision-prone (we're not merely
+    checksumming for transmission integrity).
+    """
+    s = json.dumps(d, separators=[",", ":"], sort_keys=True)
+    b = s.encode("utf-8")
+    return Version(hashlib.blake2s(b).hexdigest())
 
 
 def save_reply(source: Source, data: dict) -> Reply:
