+++ securedrop/tests/test_journalist_api2.py
@@ -8,7 +8,7 @@ import pytest
 from flask import url_for
 from flask_sqlalchemy import get_debug_queries
 from journalist_app import api2
-from journalist_app.api2 import json_version
+from journalist_app.api2.shared import json_version
 from journalist_app.api2.types import Event, EventType, ItemTarget, SourceTarget
 from models import Reply, Source, SourceStar, Submission, db
 from sqlalchemy.orm.exc import MultipleResultsFound
